+++ soccer/src/soccer/planning/planner_node.cpp
@@ -37,7 +37,7 @@ PlannerNode::PlannerNode()
 
     // set up PlannerForRobot objects
     robots_planners_.reserve(kNumShells);
-    for (int i = 0; i < kNumShells; i++) {
+    for (unsigned long i = 0; i < kNumShells; i++) {
         auto planner =
             std::make_shared<PlannerForRobot>(i, this, &robot_trajectories_, &shared_state_);
         robots_planners_.emplace_back(std::move(planner));
@@ -180,7 +180,7 @@ PlanRequest PlannerForRobot::make_request(const RobotIntent& intent) {
     const auto robot_trajectories_hold = robot_trajectories_->get();
     std::array<const Trajectory*, kNumShells> planned_trajectories = {};
 
-    for (int i = 0; i < kNumShells; i++) {
+    for (unsigned long i = 0; i < kNumShells; i++) {
         // TODO(Kevin): check that priority works (seems like
         // robot_trajectories_ is passed on init, when no planning has occured
         // yet)
