+++ soccer/src/soccer/radio/sim_radio.cpp
@@ -73,7 +73,7 @@ SimRadio::SimRadio(bool blue_team)
       blue_team_(blue_team),
       socket_(io_service_, ip::udp::endpoint(ip::udp::v4(), blue_team ? kSimBlueStatusPort
                                                                       : kSimYellowStatusPort)) {
-    for (int i = 0; i < kNumShells; ++i) {
+    for (unsigned long i = 0; i < kNumShells; ++i) {
         last_sent_diff_.emplace_back(RJ::now());
     }
 
@@ -155,7 +155,7 @@ void SimRadio::handle_receive(const std::string& data) {
 
     packet.ParseFromString(data);
 
-    for (size_t pkt_idx = 0; pkt_idx < packet.feedback_size(); pkt_idx++) {
+    for (int pkt_idx = 0; pkt_idx < packet.feedback_size(); pkt_idx++) {
         rj_msgs::msg::RobotStatus status_ros;
         RobotStatus status;
         const RobotFeedback& sim_status = packet.feedback(pkt_idx);
@@ -169,7 +169,7 @@ void SimRadio::handle_receive(const std::string& data) {
 void SimRadio::stop_robots() {
     RobotControl sim_packet;
 
-    for (int shell = 0; shell < kNumShells; shell++) {
+    for (unsigned long shell = 0; shell < kNumShells; shell++) {
         auto* sim_robot = sim_packet.add_robot_commands();
         sim_robot->set_id(shell);
         auto* local_velocity = sim_robot->mutable_move_command()->mutable_local_velocity();
